requirements.txt
openai
streamlit
pyyaml
ollama